Add unit test and fix some logic

+++ src/types/Common.sol
@@ -1,10 +1,10 @@
 // SPDX-License-Identifier: UNLICENSED
 pragma solidity ^0.8.24;
 
-struct Attestation{
+struct Attestation {
     bytes32 attestationId;
     address recipient;
     string exchange;
     uint32 baseValue;
-    uint64 timestamp;
+    uint256 timestamp;
 }
